requirements.txt
PyMuPDF>=1.23.0
ebooklib>=0.18
Pillow>=10.0.0
rarfile>=4.1
